feat: Use input detection in the right place
